Create Functions
 tech company has hired you to analyze data for their social media platform. The social media platform
collects user engagement data (data for likes, comments, shares, etc.) for the users and the posts, and
you are to develop these five functions below to organize and make sense of the data.
def filter_popular(reacts_2D, names, threshold):
Description: You are provided a two-dimensional list of engagement data counts for some users, a
one-dimensional list of those user names, and an integer threshold to determine the popular users. A
user is popular if the total engagement count for that user is equal or more than the threshold. Your
function should return a one-dimensional list of users who are popular.
Parameters:
•
reacts_2D - A 2D list. Each item is a list that contains integers representing engagement data
counts for a user.
•
names - A 1D list. Each item is a string representing a user name.
•
threshold - An integer representing the threshold for determining the popular users.
Assumptions:
•
reacts_2D and names lists are in the same order and of same length. This means first item of
reacts_2D has engagement data counts for the first user in names, second item of reacts_2D
has engagement data counts for the second user in names, and so on.
•
Each name in the names list is unique.
•
Items in reacts_2D can be empty lists.
Return value: A 1D list. Each item is a user name string from names indicating a popular user.